+++ src/constants.ts
@@ -1,0 +1,4 @@
+import { name, publisher } from '../package.json';
+
+export const EXTENSION_NAME = 'qtForPython';
+export const EXTENSION_ID = `${publisher}.${name}`;