+++ src/extension.ts
@@ -1,10 +1,7 @@
 import * as vscode from 'vscode';
-import { name, publisher } from '../package.json';
 import { commands } from './commands';
+import { EXTENSION_NAME } from './constants';
 import { showErrorMessage } from './message';
-
-export const extensionId = `${publisher}.${name}`;
-export const extensionName = 'qtForPython';
 
 export function activate(context: vscode.ExtensionContext) {
   registerCommands(context);
@@ -14,7 +11,7 @@ function registerCommands(context: vscode.ExtensionContext) {
   return commands.map(command =>
     context.subscriptions.push(
       vscode.commands.registerCommand(
-        `${extensionName}.${command.name}`,
+        `${EXTENSION_NAME}.${command.name}`,
         async (...args) => {
           try {
             return await command.callback(...args);