+++ src/tools/tool.ts
@@ -12,6 +12,8 @@ export class Tool {
     const toolPath = vscode.workspace
       .getConfiguration(pathConfigBaseName)
       .get<string>(this.name);
+    // eslint-disable-next-line no-console
+    console.log(pathConfigBaseName, this.name, toolPath);
     if (toolPath) return toolPath;
 
     return python('find_tool', [this.filename]);