+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.0 - 2021-??-??
+
+- [ ] Smart Python path.
+- [ ] Smart tools discover.
+- [ ] Auto UI compile.
+- [ ] Auto RCC compile.
+- [x] Support Linux.
+- [ ] Support Windows.
+- [x] Support PySide6.
+- [ ] Support PyQt6.
+- [ ] Support PySide2.
+- [ ] Support PyQt5.
+
 ## 0.6.2 - 2021-04-15
 
 ### Updated